[Fix] Fix talker2code2wav dimension slicing and async chunk test params
- Fix codec_codes slicing in talker2code2wav: `[-seq_len:]` sliced rows
  instead of columns on [num_quantizers, seq_len+1] tensor; changed to
  `[:, -seq_len:]` to correctly slice the time dimension
- Fix 3 async chunk tests that forgot to pass is_finished=True to the
  function call (the parameter was set on the mock request but not
  passed as a function argument)
- Merge test_qwen3_omni_pd.py into test_qwen3_omni.py with separate
  pd_test_params using the PD stage config YAML
